requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
yt-dlp==2023.11.16
opencv-python==4.8.1.78
Pillow==10.1.0
requests==2.31.0
Werkzeug==3.0.1
numpy<2.0
